requirements.txt
streamlit 
google-generativeai
python-dotenv
langchain
chromadb  # for vector embeddings
PyPDF2